refactor: update localization box properties to use decimal types and enhance image rendering logic

+++ src/Application/OrganisationScreenings/Interfaces/IOrganisationScreeningPdfService.cs
@@ -10,10 +10,10 @@ public sealed record AiFindingDetail
 
 public sealed record AiLocalizationBox
 {
-    public int X { get; init; }
-    public int Y { get; init; }
-    public int Width { get; init; }
-    public int Height { get; init; }
+    public decimal X { get; init; }
+    public decimal Y { get; init; }
+    public decimal Width { get; init; }
+    public decimal Height { get; init; }
     public decimal? Confidence { get; init; }
 }
